Further progress on collision detection
In the middle of adding xyt a priori collision detection. Follow comments in detectCollisions for more info.

Managed to get intersection points of an arbitrary number of lines and restrict the set of intersection points to those that will be on an 800x800 window.

This also means no more resizing! Only 800x800 allowed.

Have to solve a weird vector math problem to progress. I guess it's not that weird, actually, and I'm pretty sure someone's already solved it. I just don't know the specific terms to search for, and a little bit of trig won't kill me.

Once t coords are added, then we have to deal with the fact that xy coords are representative of rectangles, and collisions happen when rectangles intersect, not just lines.

I'm not exactly sure how to tackle that problem.