More collision detection?
The math is right; the java double precision is not. I think.

Theoretically, I am checking each line against every other line not-itself to see if they are skew lines. If not, then I check to see if they intersect. However, the normal line to the two non-skew-lined plane is slightly off, which means the two lines, when transformed from Z3 to N2, still have a tiny component in the normal-line direction.

I wouldn't mind it so much, but it's not all that tiny of a component, and one error in one direction means that there's errors in other directions, and since all of these vectors are generally normallized, they're already <=1, so precision matters more for these intermediary steps than one might hope.

I could kludge it and force the t in (xyt) to be zero, I guess.

Is it progress when you take a step backwards and forwards at the same time?